backend/requirements.txt
flask
requests
python-dotenv
googlemaps
firebase-admin
scikit-learn
pandas
gunicorn
